requirements.txt
peft
bitsandbytes
trl
datasets
torch
transformers
tiktoken
deepspeed
accelerate
pandas
numpy
pillow
python-dotenv
psutil
huggingface_hub